Enemies attack, animation and health system.
I added enemy movement system, I did some changes in the priority in order to give the highest priority to the player by giving to all the character priority = 1 in the nav mesh agent component. Also I added Health system in enemies and death animation which is depended to health of the enemy.