Region script which handles per region stats and it unloads/loads the gameobject
I added a region script in order to handle some data like kills per region, correct answers and so on. This feature has been added because it is necessary to construct the tasks per region. Also this script gives to the game an better performance because the units get loaded and unloaded per region depending to user's location.